# swagger.yaml
...
swagger: "2.0"
info:
  title: "Merlin"
  description: "API Guide for accessing Merlin's model deployment functionalities"
  version: "0.6.0"
host: "localhost:3000"
basePath: "/v1"
...
schemes:
  - "http"
paths:
  "/environments":
    get:
      tags: ["environment"]
      summary: "List available environment"
      description: "Environment can be filtered by optional `name` parameter"
      parameters:
        - in: "query"
          name: "name"
          required: false
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Environment"
  "/projects":
    get:
      tags: ["project"]
      summary: "List existing projects"
      description: "Projects can be filtered by optional `name` parameter"
      parameters:
        - in: "query"
          name: "name"
          required: false
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Project"
    post:
      tags: ["project"]
      summary: "Create new project"
      parameters:
        - in: "body"
          name: "body"
          description: "Project object that has to be added"
          required: true
          schema:
            $ref: "#/definitions/Project"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Project"
        400:
          description: "Invalid request format"
        409:
          description: "Project with the same name already exists"
...
  "/logs":
    get:
      tags: ["log"]
      summary: "Retrieve log from a container"
      parameters:
        - in: "query"
          name: "name"
          type: "string"
          required: true
        - in: "query"
          name: "pod_name"
          type: "string"
          required: true
        - in: "query"
          name: "namespace"
          type: "string"
          required: true
        - in: "query"
          name: "cluster"
          type: "string"
          required: true
        - in: "query"
          name: "follow"
          type: "string"
          required: false
        - in: "query"
          name: "limit_bytes"
          type: "string"
          required: false
        - in: "query"
          name: "pretty"
          type: "string"
          required: false
        - in: "query"
          name: "previous"
          type: "string"
          required: false
        - in: "query"
          name: "since_seconds"
          type: "string"
          required: false
        - in: "query"
          name: "since_time"
          type: "string"
          required: false
        - in: "query"
          name: "tail_lines"
          type: "string"
          required: false
        - in: "query"
          name: "timestamps"
          type: "string"
          required: false
      responses:
        200:
          description: "OK"
          content:
            text/plain:
              schema:
                type: "string"
...
definitions:
...
  Environment:
    type: "object"
    required:
      - name
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      cluster:
        type: "string"
      is_default:
        type: "boolean"
      region:
        type: "string"
      gcp_project:
        type: "string"
      default_resource_request:
        $ref: "#/definitions/ResourceRequest"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
  Project:
    type: "object"
    required:
      - name
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      mlflow_tracking_url:
        type: "string"
      administrators:
        type: "array"
        items:
          type: "string"
      readers:
        type: "array"
        items:
          type: "string"
      team:
        type: "string"
      stream:
        type: "string"
      labels:
        type: "array"
        items:
          $ref: "#/definitions/Label"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
...
  ResourceRequest:
    type: "object"
    properties:
      min_replica:
        type: "integer"
      max_replica:
        type: "integer"
      cpu_request:
        type: "string"
      memory_request:
        type: "string"
...
  Label:
    type: "object"
    properties:
      key:
        type: "string"
      value:
        type: "string"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []